requirements.txt
jupyter
pandas
matplotlib
seaborn
sqlalchemy
pymysql
pymongo
python-dotenv
anthropic
